+++ src/Resources/contao/dca/tl_settings.php
@@ -14,33 +14,37 @@
 /**
  * palettes
  */
-$GLOBALS['TL_DCA']['tl_settings']['palettes']['default'] .= ';{volunteeringlist_legend:hide},volunteeringlist_picWidth,volunteeringlist_picHeight,volunteeringlist_css';
+$GLOBALS['TL_DCA']['tl_settings']['palettes']['default'] .= ';{volunteeringlist_legend:hide},volunteeringlist_defaultImage,volunteeringlist_imageSize,volunteeringlist_css';
 
 /**
  * fields
  */
 
-$GLOBALS['TL_DCA']['tl_settings']['fields']['volunteeringlist_picWidth'] = array
+$GLOBALS['TL_DCA']['tl_settings']['fields']['volunteeringlist_defaultImage'] = array
 (
-	'label'         => &$GLOBALS['TL_LANG']['tl_settings']['volunteeringlist_picWidth'],
-	'inputType'     => 'text',
-	'eval'          => array
+	'label'                   => &$GLOBALS['TL_LANG']['tl_settings']['volunteeringlist_defaultImage'],
+	'inputType'               => 'fileTree',
+	'eval'                    => array
 	(
-		'tl_class'  => 'w50',
-		'rgxp'      => 'digit'
+		'filesOnly'           => true,
+		'fieldType'           => 'radio',
+		'tl_class'            => 'w50'
 	)
 );
 
-$GLOBALS['TL_DCA']['tl_settings']['fields']['volunteeringlist_picHeight'] = array
+$GLOBALS['TL_DCA']['tl_settings']['fields']['volunteeringlist_imageSize'] = array
 (
-	'label'         => &$GLOBALS['TL_LANG']['tl_settings']['volunteeringlist_picHeight'],
-	'inputType'     => 'text',
-	'eval'          => array
-	(
-		'tl_class'  => 'w50',
-		'rgxp'      => 'digit'
-	)
-);
+	'label'                   => &$GLOBALS['TL_LANG']['tl_settings']['volunteeringlist_imageSize'],
+	'exclude'                 => true,
+	'inputType'               => 'imageSize',
+	'reference'               => &$GLOBALS['TL_LANG']['MSC'],
+	'eval'                    => array('rgxp'=>'natural', 'includeBlankOption'=>true, 'nospace'=>true, 'helpwizard'=>true, 'tl_class'=>'w50'),
+	'options_callback' => static function ()
+	{
+		return System::getContainer()->get('contao.image.image_sizes')->getOptionsForUser(BackendUser::getInstance());
+	},
+	'sql'                     => "varchar(255) NOT NULL default ''"
+); 
 
 $GLOBALS['TL_DCA']['tl_settings']['fields']['volunteeringlist_css'] = array
 (
@@ -48,6 +52,6 @@ $GLOBALS['TL_DCA']['tl_settings']['fields']['volunteeringlist_css'] = array
 	'inputType'     => 'checkbox',
 	'eval'          => array
 	(
-		'tl_class'  => 'w50',
+		'tl_class'  => 'w50 clr',
 	)
 );